fix(logistics): lock scroll for the export options dialog too
回報的是**另一個**遮罩 —— 匯出選項對話框,不是上一支修的進度遮罩。
同一個根因,而上一支只修了其中一個。

## 根因與為何截圖看起來變短

`fixed inset-0` 只覆蓋**視口**而非整份文件。遮罩開啟時頁面仍可捲動,
捲下去就看到遮罩之外的內容;整頁截圖也只會擷到一個視口的遮罩,下方全部外露。

兩個不同的遮罩都在截圖裡呈現同樣的「變短」,這其實指向後者 ——
`bg-gray-900/50 fixed inset-0` 的對話框背景在正常瀏覽下確實覆蓋整個視口。
但兩種現象的修法相同,故不需要先分辨是哪一種。

**鎖定捲動同時解決兩者**:body 設 overflow:hidden 後文件高度被夾到視口,
既沒有可捲出的內容,整頁截圖也就等於一個視口。

## 抽成 useScrollLock

上一支把鎖定邏輯直接寫在 page.tsx 的 useEffect 裡,現在第二處也需要,
再複製一份就是本輪反覆在修的那個錯誤。抽為 src/hooks/use_scroll_lock.ts,
兩處共用。已確認全庫僅此一份實作。

hook 內的兩個細節:
- 於 documentElement 與 body 同時設定 —— 哪一個是實際的捲動容器取決於
  瀏覽器與樣式,只設一個在部分情況下無效
- 還原時寫回**進入前的字面值**而非清空 —— 該元素可能本來就有 inline style,
  清空會把別人的設定一併抹掉

鎖定條件改為 `isExporting || exportModalTarget !== null`,涵蓋兩個階段。

對話框遮罩另補 `h-dvh w-screen`,與進度遮罩一致:本頁匯出流程會把 viewport meta
改為 `width=1024`,而部分瀏覽器的 `inset-0` 以 layout viewport 為準,
兩者不一致時高度會短少。

## 範圍說明

全庫共 118 處 `fixed inset-0`,其中多數是其他功能的對話框(薪資計算機、
AI 諮詢室、附件預覽等),同樣沒有捲動鎖定。本輪只處理回報的兩處 ——
hook 已就位,其餘可逐步套用,但一次改 118 處無法驗證。

## 驗證

- tsc 與 develop 差值 0(189=189)、ESLint exit 0、Prettier clean
- 確認 useScrollLock 全庫僅一份實作,無複本

未驗證:仍無可用瀏覽器,無法確認實際捲動與截圖行為。若部署後仍如此,
請告知是「捲動時外露」或「僅整頁截圖如此」—— 兩者指向不同的下一步。
