+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "isunfa",
-  "version": "0.12.0+196",
+  "version": "0.12.0+197",
   "private": true,
   "type": "module",
   "engines": {
